Updating readme, adding function documentation and making a few more improvements
