+++ export_test.go
@@ -2,3 +2,5 @@ package go_tfidf
 
 var NormalizedTermFrequency = normalizedTermFrequency
 var InverseDocumentFrequency = inverseDocumentFrequency
+var RemoveDuplicates = removeDuplicates
+var StringArrayContainsWord = stringArrayContainsWord
