+++ similarity/export_test.go
@@ -1,0 +1,4 @@
+package similarity
+
+var VectorMagnitude = vectorMagnitude
+var ProductDot = productDot
