+++ similarity/similarity.go
@@ -1,3 +1,4 @@
+// Package similarity provides similarity algorithm implementations
 package similarity
 
 import (
@@ -6,6 +7,9 @@ import (
 	"strings"
 )
 
+// CalculateSimilarities receives the query individual tf-idf, query tf-idf computed for every documents in *TfIdf object, and a string with the desired similarity function to be used.
+// The default similarity function is Cosine.
+// If the TfIdf vectors have different lengths, returns an error.
 func CalculateSimilarities(queryTfIdf []float64, queryTfIdfDocuments [][]float64, similarityFunction string) ([]float64, error) {
 	var err error
 	similarities := make([]float64, 0)
@@ -21,7 +25,7 @@ func CalculateSimilarities(queryTfIdf []float64, queryTfIdfDocuments [][]float64
 		similarity := similarityFunctionToBeCalled.(func([]float64, []float64) float64)(queryTfIdf, docQueryTfidf)
 		if similarity < 0.0 {
 			similarities = make([]float64, 0)
-			err = errors.New("Vectors have different lengths")
+			err = errors.New("vectors have different lengths")
 			break
 		}
 		similarities = append(similarities, similarity)
@@ -30,20 +34,23 @@ func CalculateSimilarities(queryTfIdf []float64, queryTfIdfDocuments [][]float64
 	return similarities, err
 }
 
+// Cosine implements the Cosine similarity algorithm.
+// Receives two vectors for computing its similarity.
+// If the vector lengths are different, the function returns -1.0, indicating error.
 func Cosine(a []float64, b []float64) float64 {
 	if len(a) != len(b) {
 		return -1.0
 	}
 
-	magnitudes := VectorMagnitude(a) * VectorMagnitude(b)
+	magnitudes := vectorMagnitude(a) * vectorMagnitude(b)
 	if magnitudes > 0.0 {
-		return ProductDot(a, b) / magnitudes
+		return productDot(a, b) / magnitudes
 	}
 
 	return 0.0
 }
 
-func VectorMagnitude(vector []float64) float64 {
+func vectorMagnitude(vector []float64) float64 {
 	squareSums := 0.0
 
 	for _, v := range vector {
@@ -53,7 +60,7 @@ func VectorMagnitude(vector []float64) float64 {
 	return math.Sqrt(squareSums)
 }
 
-func ProductDot(vectorA []float64, vectorB []float64) float64 {
+func productDot(vectorA []float64, vectorB []float64) float64 {
 	if len(vectorA) != len(vectorB) {
 		return -1.0
 	}
